requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==4.0.0
click==8.1.7
ecdsa==0.18.0
fastapi==0.103.0
fastapi-pagination==0.12.9
greenlet==2.0.2
h11==0.14.0
idna==3.4
ingredient-parser-nlp==0.1.0b4
joblib==1.3.2
Mako==1.2.4
MarkupSafe==2.1.3
nltk==3.8.1
passlib==1.7.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
python-crfsuite==0.9.9
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
regex==2023.8.8
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
uvicorn==0.23.2
